backend/requirements.txt
flask
flask-cors
requests
pyyaml
requests-oauthlib
oauthlib
pytz
OTXv2
pymisp